+++ src/operation/end-when.ts
@@ -1,0 +1,6 @@
+import { TOperation } from './types';
+
+export const endWhen: TOperation<{}> = function (operationData: {}) {
+  delete this.skipNextOperation;
+  return operationData;
+};