+++ src/operation/metadata/end-when.ts
@@ -1,0 +1,9 @@
+import { IOperationMetadata } from './types';
+
+function endWhen(): IOperationMetadata<never> {
+  return {
+    description: 'Ends the current when block',
+  };
+}
+
+export default endWhen;